+++ src/Raft/Action.hs
@@ -1,5 +1,7 @@
 {-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE StandaloneDeriving #-}
+{-# LANGUAGE UndecidableInstances #-}
 
 module Raft.Action where
 
@@ -9,6 +11,7 @@ import Raft.Client
 import Raft.RPC
 import Raft.Log
 import Raft.Event
+import Raft.StateMachine
 import Raft.Types
 
 data Action sm v
@@ -16,10 +19,11 @@ data Action sm v
   | SendRPCs (Map NodeId (SendRPCAction v)) -- ^ Send a unique message to specific nodes in parallel
   | BroadcastRPC NodeIds (SendRPCAction v) -- ^ Broadcast the same message to all nodes
   | AppendLogEntries (Entries v) -- ^ Append entries to the replicated log
-  | RespondToClient ClientId (ClientRespSpec sm) -- ^ Respond to client after a client request
+  | RespondToClient ClientId (ClientRespSpec sm v) -- ^ Respond to client after a client request
   | ResetTimeoutTimer Timeout -- ^ Reset a timeout timer
   | UpdateClientReqCacheFrom Index -- ^ Update the client request cache from the given index onward
-  deriving (Show)
+
+deriving instance (Show sm, Show v, Show (RaftStateMachinePureError sm v)) => Show (Action sm v)
 
 data SendRPCAction v
   = SendAppendEntriesRPC (AppendEntriesData v)